requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-jose
passlib
transformers
torch
